+++ components/tutors-ui/index.js
@@ -2,3 +2,5 @@ export { default as Image } from './lib/Atoms/Image/Image.svelte';
 
 export { default as Card } from './lib/Molecules/Cards/Card.svelte';
 export { default as StudentCard } from './lib/Molecules/Cards/StudentCard.svelte';
+
+export { default as CardDeck } from './lib/Organisms/CardDeck/CardDeck.svelte';
